requirements.txt
asgiref==3.4.1
beautifulsoup4==4.10.0
confusable-homoglyphs==3.2.0
Django==3.2.8
django-bootstrap-v5==1.0.6
importlib-metadata==3.10.1
python-decouple==3.5
pytz==2021.3
soupsieve==2.2.1
sqlparse==0.4.2
zipp==3.6.0
